Update README.md
no issues in the branch so i put the issue here.

I can´t log in in the app, also it feels as poorly maintained and tested, as it forces you to manually install new versions instead of doing it automatically. The new version comes with critical bugs and you can´t use the old one.